requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
uvloop==0.19.0

# API & Data Validation
pydantic==2.5.0
python-multipart==0.0.6

# Database
supabase==1.2.0
psycopg2-binary==2.9.9
pandas==2.1.3

# Environment & Configuration
python-dotenv==1.0.0

# Machine Learning & Data Processing
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2

# PDF Generation & Reports
reportlab==4.0.7

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Rate Limiting
slowapi==0.1.9

# Async Support
aiofiles==23.2.1
httpx==0.25.2

# HTTP Requests
requests==2.31.0

# Logging & Monitoring
python-json-logger==2.0.7

# Utilities
pytz==2023.3
python-slugify==8.0.1

# Development (Optional - Comment out in production)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# black==23.12.0
# flake8==6.1.0

